requirements.txt
asgiref==3.2.10
autopep8==1.4.4
cassandra-driver==3.24.0
certifi==2019.3.9
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
Flask==1.1.2
geomet==0.2.1.post1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
schedule==0.6.0
six==1.15.0
sqlparse==0.2.4
stripe==2.27.0
Unipath==1.1
urllib3==1.24.2
Werkzeug==1.0.1
whitenoise==5.2.0
